backend/chat_service/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-SocketIO
Flask-JWT-Extended
eventlet